Error handling after increasing the array index
After increasing the array index, an error handling must be included, because if the input is extended later again, we can look for the error again. A handled error is easier to find and fix than a buffer overwrite error.
Table index should start from 0, because it indicates better that the table is empty.

Signed-off-by: Janos Varga <113785741+vargajb@users.noreply.github.com>